rework cpu utils so it will be consumable by functional tests
