+++ pkg/virt-launcher/virtwrap/manager.go
@@ -159,7 +159,7 @@ func (l *LibvirtDomainManager) SyncVMI(vmi *v1.VirtualMachineInstance, useEmulat
 
 	// Adjust guest vcpu config. Currenty will handle vCPUs to pCPUs pinning
 	if vmi.Spec.Domain.CPU != nil && vmi.Spec.Domain.CPU.DedicatedCPUPlacement {
-		if err := formatDomainCPUTune(vmi, domain); err != nil {
+		if err := util.FormatDomainCPUTune(vmi, domain); err != nil {
 			logger.Reason(err).Error("failed to format domain cputune.")
 			return nil, err
 		}
